nps-seasonal-model/requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
fastapi>=0.110
uvicorn[standard]>=0.29
streamlit>=1.35
plotly>=5.20
pytest>=8.0
httpx>=0.27
aiosqlite>=0.19.0
apscheduler>=3.10.4
twilio>=9.0.0
sendgrid>=6.11.0
aiosmtplib>=3.0.0
python-dotenv>=1.0.0
pytest-asyncio>=0.23.0
pydantic>=2.0
